Proxy modules moved inside cassandra-bigtable-proxy directory, fixed PR comments
