+++ cassandra-bigtable-proxy/collectiondecoder/collectiondecoder.go
@@ -13,6 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
+
 package collectiondecoder
 
 import (
