+++ cassandra-bigtable-proxy/collectiondecoder/collectiondecoder_test.go
